Leon/dashboard (#5)
* add @types/uuid npm module

* hide v2 homepage view
